asl-current-142-bld43

+++ asmmac.c
@@ -416,7 +416,9 @@ void PrintMacroList(void)
      WrLstLine(Context.OneS);
    }
    WrLstLine("");
-   sprintf(Context.OneS, "%7d%s", Context.Sum, getmessage((Context.Sum == 1) ? Num_ListMacSumMsg : Num_ListMacSumsMsg));
+   sprintf(Context.OneS, "%7lu%s",
+           (unsigned long)Context.Sum,
+           getmessage((Context.Sum == 1) ? Num_ListMacSumMsg : Num_ListMacSumsMsg));
    WrLstLine(Context.OneS);
    WrLstLine("");
 }
